ninja: Accept log file version 6
The .ninja_log file version was incremented from 5 to 6 in the 1.12.0 official
Ninja release, but there was no actual change to the format in that release.
In the 1.9.0 release, a change was made to the log file format affecting the
timestamp field (which perfetto ignores), but the ninja log file version was
not incremented at the time. The increment to 6 in 1.12.0 was made to
restore the ability to know the timestamp format for version 6 and later.
See https://github.com/ninja-build/ninja/pull/2240

The differences are irrelevant for perfetto. The format is the same, only the
meaning of the timestamp field is affected, and perfetto ignores that field
anyway. Therefore, update the header version check so that perfetto
accepts version 6 as well as version 5 ninja log files.